Expose DefualtConnectionListener as a top level class
